Mosip 14239 kyc photo issue develop 4 (#429)
* Added unit test for kyc photo. Test fix.

* Reverted sample data in common

* Added ZK controller

* Test fix

* Ignoring test as it fails in github workflow but locally builds

* Fix to get bio for photo attribute

* Removed ZK controller

* Updated kernel core version to avoid keymanager dependency runtime error

* Removed conflicting reference to older version of kernel keymanager

* Removed conflicting reference to older version of kernel keymanager

* Fixed issue

* Corrected id column type in uin hash salt table

+++ authentication/authentication-common/src/main/java/io/mosip/authentication/common/service/repository/UinHashSaltRepo.java
@@ -13,7 +13,7 @@ import io.mosip.authentication.common.service.entity.UinHashSalt;
  * @author Arun Bose S
  */
 @Repository
-public interface UinHashSaltRepo extends JpaRepository<UinHashSalt, Long> {
+public interface UinHashSaltRepo extends JpaRepository<UinHashSalt, Integer> {
 	
 	/**
 	 * The Query to retrieve salt by passing id as parameter.
@@ -22,5 +22,5 @@ public interface UinHashSaltRepo extends JpaRepository<UinHashSalt, Long> {
 	 * @return String salt
 	 */
 	@Query("select salt from UinHashSalt where id = :id")
-	public String retrieveSaltById(@Param("id") Long id);
+	public String retrieveSaltById(@Param("id") Integer id);
 }
